+++ include/index/eval/ir_eval.h
@@ -30,7 +30,7 @@ namespace index
 class ir_eval
 {
   public:
-    using result_type = std::vector<std::pair<doc_id, double>>;
+    using result_type = std::vector<std::pair<doc_id, float>>;
 
     /**
      * @param config_file Path to cpptoml configuration file
