+++ src/index/ranker/absolute_discount.cpp
@@ -15,22 +15,22 @@ namespace index
 
 const std::string absolute_discount::id = "absolute-discount";
 
-absolute_discount::absolute_discount(double delta) : delta_{delta}
+absolute_discount::absolute_discount(float delta) : delta_{delta}
 {
     /* nothing */
 }
 
-double absolute_discount::smoothed_prob(const score_data& sd) const
+float absolute_discount::smoothed_prob(const score_data& sd) const
 {
-    double pc = static_cast<double>(sd.corpus_term_count) / sd.total_terms;
-    double numerator = std::max<double>(sd.doc_term_count - delta_, 0);
-    double denominator = sd.doc_size;
+    float pc = static_cast<float>(sd.corpus_term_count) / sd.total_terms;
+    float numerator = std::max<float>(sd.doc_term_count - delta_, 0);
+    float denominator = sd.doc_size;
     return numerator / denominator + doc_constant(sd) * pc;
 }
 
-double absolute_discount::doc_constant(const score_data& sd) const
+float absolute_discount::doc_constant(const score_data& sd) const
 {
-    double unique = sd.doc_unique_terms;
+    float unique = sd.doc_unique_terms;
     return delta_ * unique / sd.doc_size;
 }
