+++ src/index/ranker/lm_ranker.cpp
@@ -16,15 +16,15 @@ namespace index
 
 const std::string language_model_ranker::id = "language-model";
 
-double language_model_ranker::score_one(const score_data& sd)
+float language_model_ranker::score_one(const score_data& sd)
 {
-    double ps = smoothed_prob(sd);
-    double pc = static_cast<double>(sd.corpus_term_count) / sd.total_terms;
+    float ps = smoothed_prob(sd);
+    float pc = static_cast<float>(sd.corpus_term_count) / sd.total_terms;
     return sd.query_term_count
            * fastapprox::fasterlog(ps / (doc_constant(sd) * pc));
 }
 
-double language_model_ranker::initial_score(const score_data& sd) const
+float language_model_ranker::initial_score(const score_data& sd) const
 {
     return sd.query.length() * fastapprox::fasterlog(doc_constant(sd));
 }
